blamediff: Fix special cases
When a file on either side of the comparison consists of only one line
the hunk header no longer contains an end number, printing "1" instead
of "1,1". Additionally, when -U0 is used the end number may be zero.
Handle these cases.

When the end number is zero git blame is run for one line unnecessarily.
I haven't bothered avoiding that.